Update stop-queue script

+++ scripts/stop-queue.sh
@@ -1,3 +1,4 @@
 #!/bin/bash
 pid=$(ps aux | grep -v "grep" | grep "php artisan" | awk '{print $2}')
-kill -9 $pid
+if [ -n "$pid" ]; then
+  kill -9 $pid
